Add an ability to route changes to different webhooks depending on prefix
Set `$wgDiscordPrefixWebhookURLs` in LocalSettings.php to be the
`prefix => hook` associative array to use.
Pages which do not match any of the prefixes would be sent to the
default webhook instead.

Only works on page creations, edits and deletions;
not on page moves, or any operations with files, users or anything
remotely esotheric.